♻️ refactor : 요소 랜더링부분이랑 분류  #199
