+++ src/components/editor/editor-container.tsx
@@ -1,12 +1,6 @@
 'use client';
 
-import {
-  ImageElement,
-  QrElement,
-  SocialElement,
-  TextElement,
-  UploadElement,
-} from '@/types/editor.type';
+import { TextElement } from '@/types/editor.type';
 import Konva from 'konva';
 import { useEffect, useMemo, useRef } from 'react';
 import { Layer, Rect, Stage, Transformer } from 'react-konva';
@@ -16,21 +10,15 @@ import {
   BASE_STAGE_HEIGHT,
   BASE_STAGE_WIDTH,
   ElEMENT_TYPE,
-  TOOLBAR_WIDTH,
 } from '@/constants/editor.constant';
 import { sideBarStore } from '@/store/editor.sidebar.store';
 import ElementToolbar from './editor-ui/element-toolbar/editor-element-toolbar';
-import UnsplashImageElement from './elements/images/element-image-canvas';
-import QrCanvasElement from './elements/qr-social/element-qr-canvas';
-import TextCanvasElement from './elements/text/element-text-canvas';
-import UploadImageElement from './elements/uploads/element-upload-canvas';
 import TextEditContent from './elements/text/text-edit-content';
-import { SwitchCase } from '../common/switch-case';
 import { handleWheel } from '@/utils/editor/editor-scale-event.util';
-import SocialCanvasElement from './elements/qr-social/element-social-canvas';
 import { useEditorStore } from '@/store/editor.store';
 import { useStageRefStore } from '@/store/editor.stage.store';
 import { useShallow } from 'zustand/react/shallow';
+import CanvasElementsRender from './canvas-elements-render';
 
 const ENABLEDANCHORS = {
   TEXT: ['middle-left', 'middle-right'],
@@ -56,7 +44,6 @@ const EditorContainer = () => {
     backgroundColor,
     backgroundColorBack,
     updateElement,
-    setToolbar,
     setSelectedElementId,
     setEditingElementId,
     setSelectedElementType,
@@ -70,17 +57,15 @@ const EditorContainer = () => {
       backgroundColor: state.backgroundColor,
       backgroundColorBack: state.backgroundColorBack,
       updateElement: state.updateElement,
-      setToolbar: state.setToolbar,
       setSelectedElementId: state.setSelectedElementId,
       setEditingElementId: state.setEditingElementId,
       setSelectedElementType: state.setSelectedElementType,
     }))
   );
-  const { zoom, isHorizontal, setSideBarStatus } = sideBarStore(
+  const { zoom, isHorizontal } = sideBarStore(
     useShallow((state) => ({
       zoom: state.zoom,
       isHorizontal: state.isHorizontal,
-      setSideBarStatus: state.setSideBarStatus,
     }))
   );
   const { setFrontStageRef, setBackStageRef } = useStageRefStore(
@@ -94,14 +79,14 @@ const EditorContainer = () => {
   const transformerRef = useRef<Konva.Transformer | null>(null);
   const shapeRefs = useRef<Record<string, Konva.Node>>({});
 
-  const visibleRef = useRef<Konva.Stage>(null);
-  const frontRef = useRef<Konva.Stage>(null);
-  const backRef = useRef<Konva.Stage>(null);
+  // const visibleRef = useRef<Konva.Stage>(null);
+  // const frontRef = useRef<Konva.Stage>(null);
+  // const backRef = useRef<Konva.Stage>(null);
 
-  useEffect(() => {
-    setFrontStageRef(frontRef);
-    setBackStageRef(backRef);
-  }, [setFrontStageRef, setBackStageRef]);
+  // useEffect(() => {
+  //   setFrontStageRef(frontRef);
+  //   setBackStageRef(backRef);
+  // }, [setFrontStageRef, setBackStageRef]);
 
   //앞면 뒷면
   const currentCanvasElements = isFront ? canvasElements : canvasBackElements;
@@ -139,52 +124,6 @@ const EditorContainer = () => {
       transformer.getLayer()?.batchDraw();
     }
   }, [selectedElementId, shapeRefs]);
-
-  /**
-   * node의 절대 위치에서 toolbar 좌표 업데이트
-   */
-  const handleUpdateToolbarNode = (node: Konva.Node) => {
-    requestAnimationFrame(() => {
-      const rect = node.getClientRect();
-      setToolbar({
-        x: rect.x + rect.width / 2 - (TOOLBAR_WIDTH * zoom) / 2,
-        y: rect.y + rect.height + 8,
-      });
-    });
-  };
-
-  /**
-   * 변환 종료 시 업데이트
-   */
-  const handleTransformEnd = (
-    id: string,
-    e: Konva.KonvaEventObject<Event>
-  ): void => {
-    const node = e.target;
-    const scaleX = node.scaleX();
-    const scaleY = node.scaleY();
-    node.scaleX(1);
-    node.scaleY(1);
-
-    updateElement(id, {
-      x: node.x(),
-      y: node.y(),
-      width: Math.max(10, node.width() * scaleX),
-      height: Math.max(10, node.height() * scaleY),
-      rotation: node.rotation(),
-    });
-
-    handleUpdateToolbarNode(node);
-  };
-
-  /**
-   * 더블 클릭 시 편집 모드 실행 핸들러
-   * @param id 요소 id
-   */
-  const handleTextDoubleClick = (id: string) => {
-    setSelectedElementId(id);
-    setEditingElementId(id);
-  };
 
   // 인라인 편집 완료 후 텍스트 업데이트
   const handleTextEditSubmit = (newText: string) => {
@@ -224,7 +163,7 @@ const EditorContainer = () => {
       }}
     >
       <Stage
-        ref={isFront ? frontRef : backRef}
+        // ref={isFront ? frontRef : backRef}
         style={{ border: '1px dashed var(--Gray-60, #878A93)' }}
         width={currentStageWidth}
         height={currentStageHeight}
@@ -248,71 +187,11 @@ const EditorContainer = () => {
             fill={currentBackgroundColor || '#ffffff'}
             listening={false}
           />
-
-          {currentCanvasElements.map((el) => {
-            // 공통 Props
-            const commonProps = {
-              onDragEnd: (id: string, node: Konva.Node) => {
-                updateElement(id, { x: node.x(), y: node.y() });
-                handleUpdateToolbarNode(node);
-              },
-              onDragMove: (node: Konva.Node) => {
-                handleUpdateToolbarNode(node);
-              },
-              onTransformEnd: handleTransformEnd,
-              onSelect: (id: string, node: Konva.Node) => {
-                setSelectedElementId(id);
-                handleUpdateToolbarNode(node);
-                setSelectedElementType(el.type);
-                setSideBarStatus(true);
-              },
-              ref: (node: Konva.Node | null) => {
-                if (node) shapeRefs.current[el.id] = node;
-              },
-            };
-
-            return (
-              <SwitchCase
-                key={el.id}
-                value={el.type}
-                cases={{
-                  [ElEMENT_TYPE.TEXT]: (
-                    <TextCanvasElement
-                      element={el as TextElement}
-                      {...commonProps}
-                      onDoubleClick={handleTextDoubleClick}
-                      editing={editingElementId === el.id}
-                    />
-                  ),
-                  [ElEMENT_TYPE.UPLOAD]: (
-                    <UploadImageElement
-                      element={el as UploadElement}
-                      {...commonProps}
-                    />
-                  ),
-                  [ElEMENT_TYPE.IMAGE]: (
-                    <UnsplashImageElement
-                      element={el as ImageElement}
-                      {...commonProps}
-                    />
-                  ),
-                  [ElEMENT_TYPE.QR]: (
-                    <QrCanvasElement
-                      element={el as QrElement}
-                      {...commonProps}
-                    />
-                  ),
-                  [ElEMENT_TYPE.SOCIAL]: (
-                    <SocialCanvasElement
-                      element={el as SocialElement}
-                      {...commonProps}
-                    />
-                  ),
-                }}
-                default={<></>}
-              />
-            );
-          })}
+          <CanvasElementsRender
+            elements={currentCanvasElements}
+            editingElementId={editingElementId}
+            shapeRefs={shapeRefs}
+          />
           <Transformer
             ref={transformerRef}
             enabledAnchors={
